refactor(OrderBook): move Book into a new module
